feat: main-service: - пакет request: создан слой репозитория для работы с БД, реализована сервисная логика и логика взаимодействия с пользователем (контролеры)
- создан пакет compilation: создан слой репозитория для работы с БД, реализована сервисная логика и логика взаимодействия с пользователем (контролеры)

+++ main-service/src/main/java/ru/yandex/practicum/mainserver/event/model/Event.java
@@ -67,7 +67,7 @@ public class Event {
     @OneToOne
     @JoinColumn(name = "location_id")
     private Location location;
-    private Integer confirmedRequests; // Количество одобренных заявок на участие в данном событии
+    private Long confirmedRequests; // Количество одобренных заявок на участие в данном событии
     private Long participantLimit;
     private Boolean requestModeration;
